refactor: move react-email to own package
