+++ apps/www/package.json
@@ -11,7 +11,6 @@
     "build": "contentlayer build && pnpm with-env next build",
     "clean": "git clean -xdf .next .turbo .contentlayer node_modules",
     "dev": "pnpm with-env next dev",
-    "email": "email dev --dir emails --port 3333",
     "format": "prettier --check \"**/*.{js,cjs,mjs,ts,tsx,md,json}\"",
     "lint": "dotenv -v SKIP_ENV_VALIDATION=1 next lint",
     "postinstall": "prisma generate",
@@ -32,6 +31,7 @@
     "@hookform/resolvers": "^3.3.2",
     "@next-auth/prisma-adapter": "^1.0.7",
     "@prisma/client": "^5.5.2",
+    "@projectx/transactional": "^0.1.0",
     "@radix-ui/react-accessible-icon": "^1.0.3",
     "@radix-ui/react-accordion": "^1.1.2",
     "@radix-ui/react-alert-dialog": "^1.0.5",
@@ -61,9 +61,6 @@
     "@radix-ui/react-toggle": "^1.0.3",
     "@radix-ui/react-toggle-group": "^1.0.4",
     "@radix-ui/react-tooltip": "^1.0.7",
-    "@react-email/button": "0.0.11",
-    "@react-email/components": "0.0.7",
-    "@react-email/html": "0.0.6",
     "@t3-oss/env-nextjs": "^0.7.1",
     "@tanstack/react-table": "^8.10.7",
     "@vercel/analytics": "^1.1.1",
@@ -91,14 +88,12 @@
     "react-day-picker": "^8.9.1",
     "react-dom": "^18.2.0",
     "react-dropzone": "^14.2.3",
-    "react-email": "1.9.5",
     "react-facebook-pixel": "^1.0.4",
     "react-hook-form": "^7.47.0",
     "react-resizable-panels": "^0.0.55",
     "react-textarea-autosize": "^8.5.3",
     "react-wrap-balancer": "^1.1.0",
     "recharts": "^2.10.3",
-    "resend": "^1.1.0",
     "sharp": "^0.32.6",
     "shiki": "^0.14.5",
     "stripe": "^14.2.0",
