+++ package.json
@@ -15,19 +15,19 @@
     "clean": "git clean -xdf node_modules dist .next",
     "clean:workspaces": "turbo clean",
     "build": "turbo build",
-    "dev": "turbo dev --parallel",
-    "dev:web": "turbo dev --parallel --filter !@projectx/db",
+    "dev": "turbo dev --parallel --filter=!react-email-client",
+    "dev:web": "turbo dev --parallel --filter !@projectx/transactional --filter=!react-email-client",
+    "dev:email": "turbo dev --parallel --filter !www --filter=!react-email-client",
     "test": "turbo test",
     "postinstall": "prisma generate",
-    "email": "email dev --dir emails --port 3333",
     "prisma:generate": "dotenv -e .env.local prisma generate",
     "prisma:migrate:dev": "dotenv -e .env.local prisma migrate dev",
     "prisma:studio": "dotenv -e .env.local prisma studio",
     "prisma:push": "dotenv -e .env.local prisma db push",
-    "format": "turbo format --continue -- --write --cache --cache-location='node_modules/.cache/.prettiercache' --ignore-path='../../.gitignore'",
-    "format:check": "turbo format --continue -- --cache --cache-location='node_modules/.cache/.prettiercache' --ignore-path='../../.gitignore'",
-    "lint": "turbo lint --continue -- --cache --cache-location 'node_modules/.cache/.eslintcache' && manypkg check",
-    "typecheck": "turbo typecheck"
+    "format": "turbo format --filter=!react-email-client --continue -- --write --cache --cache-location='node_modules/.cache/.prettiercache' --ignore-path='../../.gitignore'",
+    "format:check": "turbo format --filter=!react-email-client --continue -- --cache --cache-location='node_modules/.cache/.prettiercache' --ignore-path='../../.gitignore'",
+    "lint": "turbo lint --filter=!react-email-client --continue -- --cache --cache-location 'node_modules/.cache/.eslintcache' && manypkg check",
+    "typecheck": "turbo typecheck --filter=!react-email-client"
   },
   "dependencies": {
     "@manypkg/cli": "^0.21.0",
