+++ packages/transactional/src/index.ts
@@ -1,5 +1,5 @@
 import { Resend } from "resend";
 
-import { env } from "@/env.mjs";
+import { env } from "./env.mjs";
 
 export const resend = new Resend(env.RESEND_API_KEY);
