+++ turbo.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://turborepo.org/schema.json",
   "globalDependencies": [
-    "**/.env"
+    "**/.env.*local"
   ],
   "pipeline": {
     "topo": {
